refactor: fix package build and remove unneeded files
- fix Dynamic build with skipLibCheck option in tsconfig.json
- replace Tailwind with inline styles
- add .npmignore
- update dependencies
- add clean and transpile scripts
