+++ .npmignore
@@ -1,0 +1,9 @@
+src/
+tsconfig.json
+eslintrc.js
+commintlint.config.js
+.prettierrc
+.prettierignore
+node_modules/
+.vscode
+.env*
