+++ package.json
@@ -1,13 +1,24 @@
 {
   "name": "login-button",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "description": "A standalone login component with an embedded Dynamic modal that functions independently of host application.",
+  "repository": "https://github.com/fleek-platform/login-button",
+  "homepage": "https://github.com/fleek-platform/login-button",
+  "keywords": [
+    "fleek",
+    "auth",
+    "dynamic"
+  ],
+  "license": "MIT",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "build": "tsc",
-    "clean": "rm -rf dist",
-    "clean:nm": "rm -rf node_modules pnpm-lock.yaml"
+    "build": "pnpm clean:dist && pnpm transpile",
+    "transpile": "pnpm exec tsc -p tsconfig.json",
+    "clean:nm": "rm -rf node_modules pnpm-lock.yaml",
+    "clean": "pnpm \"/^clean:.*/\"",
+    "clean:dist": "rm -rf dist",
+    "clean:node": "rm -rf node_modules"
   },
   "dependencies": {
     "@dynamic-labs/ethereum": "3.9.2",
@@ -16,26 +27,19 @@
     "@fleek-platform/utils-token": "^0.2.2",
     "cookies-next": "^2.1.1",
     "luxon": "^3.5.0",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "viem": "^2.21.55"
   },
   "devDependencies": {
     "@types/luxon": "^3.4.2",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.9",
+    "eslint": "^8.57.0",
     "eslint-config-next": "14.2.7",
     "pino-pretty": "^13.0.0",
-    "postcss": "^8",
     "prettier": "^2.8.8",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fleek-platform/login-button"
-  },
-  "license": "MIT"
+    "typescript": "5.6.2"
+  }
 }
