+++ tsconfig.json
@@ -1,11 +1,13 @@
 {
   "compilerOptions": {
-    "target": "ES2015",
+    "target": "es2020",
     "module": "CommonJS",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": ["es2020", "dom"],
+    // only way to build @dynamic-labs/sdk-react-core
+    // https://github.com/reown-com/appkit/issues/1211
+    "skipLibCheck": true,
     "jsx": "preserve",
     "declaration": true,
-    "emitDeclarationOnly": true,
     "outDir": "dist",
     "strict": true,
     "esModuleInterop": true,
